// app/admin/products/new/types/productForm.types.ts
export interface ProductFormData {
  name: string;
  nameAr: string;
  shortDescription: string;
  shortDescriptionAr: string;
  description: string;
  descriptionAr: string;
  basePrice: string;
  discountType: string;
  discountValue: string;
  currency: string;
  mainImage: string;
  imageGallery: string[];
  video: string;
  stock: string;
  sku: string;
  category: string;
  categoryAr: string;
  subCategory: string;
  subCategoryAr: string;
  brand: string;
  brandAr: string;
  tags: string;
  tagsAr: string;
  sizes: string;
  colors: string;
  material: string;
  weight: string;
  length: string;
  width: string;
  height: string;
  warranty: string;
  returnPolicy: string;
  specifications: string;
  specificationsAr: string;
  adminReviewRating: string;
  adminReviewComment: string;
}

export interface FormSectionProps {
  formData: ProductFormData;
  handleChange: (e: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement | HTMLSelectElement>) => void;
}